+++ service/src/main/java/org/databiosphere/workspacedataservice/dataimport/AvroRecordConverter.java
@@ -15,7 +15,7 @@ import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericEnumSymbol;
 import org.apache.avro.generic.GenericFixed;
 import org.apache.avro.generic.GenericRecord;
-import org.databiosphere.workspacedataservice.service.TwoPassStreamingWriteHandler;
+import org.databiosphere.workspacedataservice.recordstream.TwoPassStreamingWriteHandler;
 import org.databiosphere.workspacedataservice.shared.model.Record;
 import org.databiosphere.workspacedataservice.shared.model.RecordAttributes;
 import org.slf4j.Logger;
@@ -53,7 +53,7 @@ public abstract class AvroRecordConverter {
    * @param genericRecord the inbound Avro GenericRecord to be converted
    * @return the Record containing base (non-relation) WDS attributes
    */
-  abstract Record convertBaseAttributes(GenericRecord genericRecord);
+  public abstract Record convertBaseAttributes(GenericRecord genericRecord);
 
   /**
    * When operating in {@see TwoPassStreamingWriteHandler.ImportMode.RELATIONS} mode, what Record -
@@ -62,7 +62,7 @@ public abstract class AvroRecordConverter {
    * @param genericRecord the inbound Avro GenericRecord to be converted
    * @return the Record containing WDS relation attributes
    */
-  abstract Record convertRelations(GenericRecord genericRecord);
+  public abstract Record convertRelations(GenericRecord genericRecord);
 
   /**
    * Extract WDS attributes from an Avro GenericRecord, optionally skipping over a set of
@@ -100,7 +100,7 @@ public abstract class AvroRecordConverter {
    * @param attribute the Avro field
    * @return the WDS attribute value
    */
-  Object convertAttributeType(Object attribute) {
+  public Object convertAttributeType(Object attribute) {
 
     if (attribute == null) {
       return null;
