+++ service/src/main/java/org/databiosphere/workspacedataservice/dataimport/pfb/PfbConfig.java
@@ -1,6 +1,7 @@
-package org.databiosphere.workspacedataservice.dataimport;
+package org.databiosphere.workspacedataservice.dataimport.pfb;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
