+++ service/src/main/java/org/databiosphere/workspacedataservice/recordstream/StreamingWriteHandler.java
@@ -1,4 +1,4 @@
-package org.databiosphere.workspacedataservice.service;
+package org.databiosphere.workspacedataservice.recordstream;
 
 import java.io.Closeable;
 import java.io.IOException;
