+++ service/src/main/java/org/databiosphere/workspacedataservice/service/ImportService.java
@@ -11,8 +11,8 @@ import java.util.UUID;
 import org.databiosphere.workspacedataservice.dao.JobDao;
 import org.databiosphere.workspacedataservice.dao.SchedulerDao;
 import org.databiosphere.workspacedataservice.dataimport.ImportJobInput;
-import org.databiosphere.workspacedataservice.dataimport.PfbSchedulable;
-import org.databiosphere.workspacedataservice.dataimport.TdrManifestSchedulable;
+import org.databiosphere.workspacedataservice.dataimport.pfb.PfbSchedulable;
+import org.databiosphere.workspacedataservice.dataimport.tdr.TdrManifestSchedulable;
 import org.databiosphere.workspacedataservice.generated.GenericJobServerModel;
 import org.databiosphere.workspacedataservice.generated.ImportRequestServerModel;
 import org.databiosphere.workspacedataservice.sam.SamDao;
