+++ service/src/test/java/org/databiosphere/workspacedataservice/service/ImportServiceTest.java
@@ -24,8 +24,8 @@ import org.broadinstitute.dsde.workbench.client.sam.api.GoogleApi;
 import org.broadinstitute.dsde.workbench.client.sam.api.ResourcesApi;
 import org.databiosphere.workspacedataservice.dao.JobDao;
 import org.databiosphere.workspacedataservice.dao.SchedulerDao;
-import org.databiosphere.workspacedataservice.dataimport.PfbQuartzJob;
-import org.databiosphere.workspacedataservice.dataimport.TdrManifestQuartzJob;
+import org.databiosphere.workspacedataservice.dataimport.pfb.PfbQuartzJob;
+import org.databiosphere.workspacedataservice.dataimport.tdr.TdrManifestQuartzJob;
 import org.databiosphere.workspacedataservice.generated.GenericJobServerModel;
 import org.databiosphere.workspacedataservice.generated.ImportRequestServerModel;
 import org.databiosphere.workspacedataservice.sam.SamClientFactory;
